Split SyntaxNodeExtensions file into Spacing extension
